+++ lib/services/squares/square_service.dart
@@ -1,0 +1,7 @@
+import 'package:chat_interface/services/chat/conversation_service.dart';
+
+class SquareContainer extends ConversationContainer {
+  final topics = <String>[];
+
+  SquareContainer(super.name);
+}