+++ lib/translations/squares.dart
@@ -1,0 +1,99 @@
+import 'package:get/get.dart';
+
+class SquareTranslations extends Translations {
+  @override
+  Map<String, Map<String, String>> get keys => {
+    //* English US
+    'en_US': {
+      // Square management
+      'squares.create': 'Create Square',
+    },
+
+    //* German
+    'de_DE': {
+      /*
+          // App
+          'app.title': 'The chat app',
+          'app.welcome': 'Danke, dass du mich auf dieser Reise begleitest!',
+          'app.build': 'Aktuelle Version: @build',
+
+          // Profile
+          'status.0': 'Offline',
+          'status.1': 'Online',
+          'status.2': 'Abwesend',
+          'status.3': 'Beschäftigt',
+          'status.message': 'Status',
+          'status.message.add': 'Status hinzufügen',
+          'profile.settings': 'Einstellungen',
+          'profile.friends': 'Freunde',
+          'profile.files': 'Dateien',
+          'profile.test': "Teste etwas (NICHT KLICKEN)",
+
+          // Friends
+          'friends': 'Freunde',
+          'friends.placeholder': 'Freunde durchsuchen (auch hinzufügen)',
+          'friends.remove': 'Freund entfernen',
+          'friend.removed': 'Du hast diese Person entfernt.',
+          'friends.add': 'Freund hinzufügen',
+          'friends.message': 'Direktnachricht starten',
+          'friends.invite_to_space': 'Zum Space einladen',
+          'friends.requests_sent': 'Gesendete Anfragen',
+          'friends.requests': 'Anfragen',
+          'request.sent': 'Anfrage erfolgreich gesendet!',
+          'friends.empty': 'Es scheint, als hättest du keine Freunde mit diesem Namen.',
+          'friends.send_request': 'Drücke einfach \'Enter\', wenn du fertig bist. Und wir finden deinen Freund für dich.',
+          'friends.example': 'Wenn du jemanden hinzufügen willst, gib bitte Benutzername ein.',
+          'request.confirm.title': 'Anfrage bestätigen',
+          'request.confirm.text':
+              'Wenn du eine Freundschaftsanfrage sendest, wird die andere Person dein Profil (Profilbild, Beschreibung, usw.) dauerhaft sehen können, es sei denn, du änderst deine Schlüssel. Bist du sicher, dass du dieser Person diese Informationen geben willst?',
+
+          // Conversations
+          'conversations.placeholder': 'Suche',
+          'conversations.hidden': "Suchst du nach etwas anderem? Du kannst '.' am Anfang deiner Suche hinzufügen, um nach versteckten Chats zu suchen.",
+          'conversations.create': 'Chat erstellen',
+          'conversations.name': 'Gruppenname',
+          'chat.message': 'Sag etwas',
+          'chat.members': '@count Mitglieder',
+          'chat.start_space': 'Privates Space starten',
+          'conversations.leave': 'Chat verlassen',
+          'conversations.leave.text': 'Bist du sicher, dass du diesen Chat verlassen willst? Du kannst nicht wieder beitreten, es sei denn, jemand lädt dich wieder ein.',
+          'chat.not.signed': 'Diese Nachricht könnte von jemand anderem gesendet worden sein oder vom Server verändert worden sein.',
+
+          // Message menu
+          'message.info': 'Info',
+          'message.copy': 'Text kopieren',
+          'message.profile': 'Profil öffnen',
+          'message.delete': 'Nachricht löschen',
+          'message.info.text': 'Diese Nachricht wurde von @sender (@senderid) um @hour:@minute am @day.@month.@year gesendet.',
+          'message.info.copy_signature': 'Signatur kopieren',
+          'message.info.copy_sender': 'Sender ID kopieren',
+          'message.info.copy_cert': 'Zertifikat kopieren',
+
+          // Conversation members
+          'chat.make_moderator': 'Zum Moderator machen',
+          'chat.remove_moderator': 'Moderator entfernen',
+          'chat.remove_admin': 'Admin entfernen',
+          'chat.make_admin': 'Zum Admin machen',
+          'chat.remove_member': 'Mitglied entfernen',
+          'chat.add_member': 'Mitglied hinzufügen',
+          'chat.admin': 'Admin',
+          'chat.moderator': 'Moderator',
+          'chat.user': 'Mitglied',
+
+          // System messages
+          'chat.rank_change.0->1': '@name wurde von @sender zum Moderator befördert.',
+          'chat.rank_change.1->2': '@name wurde von @sender zum Admin befördert.',
+          'chat.rank_change.1->0': '@name wurde von @sender zu einem normalen Mitglied degradiert.',
+          'chat.rank_change.2->1': '@name wurde von @sender zu einem Moderator degradiert.',
+          'chat.token_change': '@name hat eine neue Einladung zum Chat erstellt.',
+          'chat.member_join': '@name ist dem Chat beigetreten.',
+          'chat.member_leave': '@name hat den Chat verlassen.',
+          'chat.new_admin': '@name ist jetzt Admin, weil der ursprüngliche Admin den Chat verlassen hat.',
+
+          // Spaces
+          'join.space': 'Space beitreten',
+          'join.space.popup': 'Manche klicken darauf ausversehen, also willst du wirklich diesem Space beitreten?',
+          */
+    },
+  };
+}